+++ app/control_plane/test/ankole/signals_gateway/actor_runtime/worker_env_test.exs
@@ -6,7 +6,6 @@ defmodule Ankole.SignalsGateway.ActorRuntime.WorkerEnvTest do
       rpc_request: 3,
       rpc_request: 4,
       rpc_response_payload!: 2,
-      rpc_error_payload!: 1,
       envelope_body_type: 1,
       envelope_body!: 2
     ]